Bugged PID is working at least.

+++ Digital/CS301_Class.cydsn/control/pid.c
@@ -68,7 +68,7 @@ void pid_compute(PIDData* data) {
         data->output = data->output_min;
     }
     else {
-        data->output = output;
+        data->output = (int32_t)output;
     }
 }
